còn chỉnh batchcode, chỉnh promotion, chỉnh commment, chỉnh áp dụng promotion, tính ship, thống kê nhập hàng,

+++ Fullstackecommercereactjswithadminsourcecode/admin/src/pages/BatchCodes/index.js
@@ -48,6 +48,8 @@ const BatchCodeList = () => {
     }, []);
 
     const deleteBatchCode = (id) => {
+        const confirmDelete = window.confirm("Are you sure you want to delete?");
+        if (!confirmDelete) return;
         setIsLoading(true);
         context.setProgress(30);
         deleteData(`/api/batchCodes/${id}`).then(() => {
